telegram_vscode_assistant/requirements.txt
python-telegram-bot==20.4
pyautogui==0.9.53
Pillow==10.0.0
openai>=1.0.0
